pyproject.toml
[project]
name = "wine-chat"
version = "0.1.0"
description = "Wine Chat is an AI-powered platform for wine lovers. Discover wine venues, tastings, sommeliers, and curated articles. Whether you're a casual enthusiast or a connoisseur, its delivers personalized recommendations to enhance your wine journey."
authors = [
    {name = "Luiz Gustavo Gavinho",email = "lgavinho@gmail.com"}
]
license = {text = "proprietary"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.60.1,<2.0.0)",
    "fontawesome-free (>=5.15.4,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
